+++ packages/@expo/cli/jest.config.js
@@ -14,13 +14,5 @@ module.exports = {
       setupFiles: ['<rootDir>/e2e/setup.ts'],
       clearMocks: true,
     },
-    {
-      displayName: require('./package').name + '-types',
-      runner: 'jest-runner-tsd',
-      testRegex: '/__typetests__/.*(test|spec)\\.[jt]sx?$',
-      rootDir: path.resolve(__dirname),
-      roots,
-      globalSetup: '<rootDir>/src/start/server/type-generation/__typetests__/generateFixtures.ts',
-    },
   ],
 };