+++ packages/expo-router/src/hooks.ts
@@ -2,7 +2,8 @@ import React from 'react';
 
 import { LocalRouteParamsContext } from './Route';
 import { store, useStoreRootState, useStoreRouteInfo } from './global-state/router-store';
-import { ExpoRouter } from '../types/expo-router';
+import { Router } from './imperative-api';
+import { RouteParams, RouteSegments, Routes, UnknownOutputParams } from './types';
 
 type SearchParams = Record<string, string | string[]>;
 export function useRootNavigationState() {
@@ -23,7 +24,7 @@ export function useNavigationContainerRef() {
   return store.navigationRef;
 }
 
-export function useRouter(): ExpoRouter.Router {
+export function useRouter(): Router {
   return React.useMemo(
     () => ({
       push: store.push,
@@ -68,8 +69,12 @@ export function useUnstableGlobalHref(): string {
  * const [first, second] = useSegments<['settings'] | ['[user]'] | ['[user]', 'followers']>()
  * ```
  */
-export function useSegments<TSegments extends string[] = string[]>(): TSegments {
-  return useStoreRouteInfo().segments as TSegments;
+export function useSegments<
+  TSegments extends Routes | RouteSegments<Routes> = Routes,
+>(): TSegments extends string ? RouteSegments<TSegments> : TSegments {
+  return useStoreRouteInfo().segments as TSegments extends string
+    ? RouteSegments<TSegments>
+    : TSegments;
 }
 
 /** @returns global selected pathname without query parameters. */
@@ -87,9 +92,17 @@ export function usePathname(): string {
  * @see `useLocalSearchParams`
  */
 export function useGlobalSearchParams<
-  TParams extends SearchParams = SearchParams,
->(): Partial<TParams> {
-  return useStoreRouteInfo().params as Partial<TParams>;
+  TParams extends SearchParams = UnknownOutputParams,
+>(): RouteParams<TParams>;
+export function useGlobalSearchParams<
+  TRoute extends Routes,
+  TParams extends SearchParams = UnknownOutputParams,
+>(): RouteParams<TRoute, TParams>;
+export function useGlobalSearchParams<
+  TParams1 extends SearchParams | Routes = UnknownOutputParams,
+  TParams2 extends SearchParams = UnknownOutputParams,
+>(): RouteParams<TParams1, TParams2> {
+  return useStoreRouteInfo().params as RouteParams<TParams1, TParams2>;
 }
 
 /**
@@ -98,10 +111,20 @@ export function useGlobalSearchParams<
  * For dynamic routes, both the route parameters and the search parameters are returned.
  *
  * To observe updates even when the invoking route is not focused, use `useGlobalSearchParams()`.
+ *
+ * @see `useGlobalSearchParams`
  */
 export function useLocalSearchParams<
-  TParams extends SearchParams = SearchParams,
->(): Partial<TParams> {
+  TParams extends SearchParams = UnknownOutputParams,
+>(): RouteParams<TParams>;
+export function useLocalSearchParams<
+  TRoute extends Routes,
+  TParams extends SearchParams = UnknownOutputParams,
+>(): RouteParams<TRoute, TParams>;
+export function useLocalSearchParams<
+  TParams1 extends SearchParams | Routes = UnknownOutputParams,
+  TParams2 extends SearchParams = UnknownOutputParams,
+>(): RouteParams<TParams1, TParams2> {
   const params = React.useContext(LocalRouteParamsContext) ?? {};
   return Object.fromEntries(
     Object.entries(params).map(([key, value]) => {
@@ -124,5 +147,5 @@ export function useLocalSearchParams<
         }
       }
     })
-  ) as Partial<TParams>;
+  ) as RouteParams<TParams1, TParams2>;
 }